Extract menu to new header file and define translations via #define
