+++ src/display.h
@@ -154,40 +154,115 @@ void displayShottimer(void) {
 }
 
 #if (ROTARY_MENU == 1) 
-void displayTempSetting(double temp, const char* name) {
+
+void displayNumericalMenuSettingWithUnit(double temp, const char* name, const char* unit) {
     u8g2.clearBuffer();
     u8g2.setCursor(0, 0);
     u8g2.print(name);
     u8g2.print(": ");
     u8g2.setCursor(0, 10);
     u8g2.print(temp, 1);
-    u8g2.print(" ");
-    u8g2.print((char)176);
-    u8g2.print("C");
+    u8g2.print(unit);
     u8g2.setCursor(0, 25);
-    u8g2.println("Press encoder to save");
+    u8g2.println(langstring_pressToSave[0]);
     u8g2.setCursor(0, 35);
-    u8g2.println("and return.");
+    u8g2.println(langstring_pressToSave[1]);
     u8g2.sendBuffer();
 }
 
 void displayToggleBackflushMessage(uint8_t mode) {
     u8g2.clearBuffer();
     u8g2.setCursor(0, 0);
-    u8g2.print("Backflush: ");
+    u8g2.print(LANGSTRING_MENU_BACKFLUSH);
     switch (mode) {
         case 1:
-            u8g2.print("on.");
+            u8g2.print(LANGSTRING_MENU_ON);
             break;
         default:
-            u8g2.print("off.");
+            u8g2.print(LANGSTRING_MENU_OFF);
             break;
     }
     u8g2.setCursor(0, 25);
-    u8g2.println("Menu will autoclose");
+    u8g2.println(langstring_autoclose[0]);
     u8g2.setCursor(0, 35);
-    u8g2.println("in two seconds.");
+    u8g2.println(langstring_autoclose[1]);
     u8g2.sendBuffer();
+}
+
+void displayMenu() {
+    char content_text[_LCDML_DISP_cols];  // save the content text of every menu element
+    LCDMenuLib2_menu *tmp;
+    uint8_t i = LCDML.MENU_getScroll();
+    uint8_t maxi = _LCDML_DISP_rows + i;
+    uint8_t n = 0;
+
+    // init vars
+    uint8_t n_max             = (LCDML.MENU_getChilds() >= _LCDML_DISP_rows) ? _LCDML_DISP_rows : (LCDML.MENU_getChilds());
+
+    uint8_t scrollbar_min     = 0;
+    uint8_t scrollbar_max     = LCDML.MENU_getChilds();
+    uint8_t scrollbar_cur_pos = LCDML.MENU_getCursorPosAbs();
+    uint8_t scroll_pos        = ((1.*n_max * _LCDML_DISP_rows) / (scrollbar_max - 1) * scrollbar_cur_pos);
+
+    // generate content
+    u8g2.firstPage();
+    do {
+        n = 0;
+        i = LCDML.MENU_getScroll();
+        // check if this element has children
+        if ((tmp = LCDML.MENU_getDisplayedObj()) != NULL) {
+        // loop to display lines
+        do {
+            // check if a menu element has a condition and if the condition be true
+            if (tmp->checkCondition()) {
+            // check the type off a menu element
+            if(tmp->checkType_menu() == true) {
+                // display normal content
+                LCDML_getContent(content_text, tmp->getID());
+                u8g2.drawStr( _LCDML_DISP_box_x0+_LCDML_DISP_font_w + _LCDML_DISP_cur_space_behind, _LCDML_DISP_box_y0 + _LCDML_DISP_font_h * (n + 1), content_text);
+            }
+            else {
+                if(tmp->checkType_dynParam()) {
+                tmp->callback(n);
+                }
+            }
+            // increment some values
+            i++;
+            n++;
+            }
+        // try to go to the next sibling and check the number of displayed rows
+        } while (((tmp = tmp->getSibling(1)) != NULL) && (i < maxi));
+        }
+
+        // set cursor
+        u8g2.drawStr( _LCDML_DISP_box_x0+_LCDML_DISP_cur_space_before, _LCDML_DISP_box_y0 + _LCDML_DISP_font_h * (LCDML.MENU_getCursorPos() + 1),  _LCDML_DISP_cursor_char);
+
+        if(_LCDML_DISP_draw_frame == 1) {
+        u8g2.drawFrame(_LCDML_DISP_box_x0, _LCDML_DISP_box_y0, (_LCDML_DISP_box_x1-_LCDML_DISP_box_x0), (_LCDML_DISP_box_y1-_LCDML_DISP_box_y0));
+        }
+
+        // display scrollbar when more content as rows available and with > 2
+        if (scrollbar_max > n_max && _LCDML_DISP_scrollbar_w > 2)
+        {
+        // set frame for scrollbar
+        u8g2.drawFrame(_LCDML_DISP_box_x1 - _LCDML_DISP_scrollbar_w, _LCDML_DISP_box_y0, _LCDML_DISP_scrollbar_w, _LCDML_DISP_box_y1-_LCDML_DISP_box_y0);
+
+        // calculate scrollbar length
+        uint8_t scrollbar_block_length = scrollbar_max - n_max;
+        scrollbar_block_length = (_LCDML_DISP_box_y1-_LCDML_DISP_box_y0) / (scrollbar_block_length + _LCDML_DISP_rows);
+
+        //set scrollbar
+        if (scrollbar_cur_pos == 0) {                                   // top position     (min)
+            u8g2.drawBox(_LCDML_DISP_box_x1 - (_LCDML_DISP_scrollbar_w-1), _LCDML_DISP_box_y0 + 1                                                     , (_LCDML_DISP_scrollbar_w-2)  , scrollbar_block_length);
+        }
+        else if (scrollbar_cur_pos == (scrollbar_max-1)) {            // bottom position  (max)
+            u8g2.drawBox(_LCDML_DISP_box_x1 - (_LCDML_DISP_scrollbar_w-1), _LCDML_DISP_box_y1 - scrollbar_block_length                                , (_LCDML_DISP_scrollbar_w-2)  , scrollbar_block_length);
+        }
+        else {                                                                // between top and bottom
+            u8g2.drawBox(_LCDML_DISP_box_x1 - (_LCDML_DISP_scrollbar_w-1), _LCDML_DISP_box_y0 + (scrollbar_block_length * scrollbar_cur_pos + 1),(_LCDML_DISP_scrollbar_w-2)  , scrollbar_block_length);
+        }
+        }
+    } while ( u8g2.nextPage() );
 }
 #endif
